Failed attempt to make full refresh option.
Commiting in case I missed something, it doesnt work because CaseNumber is a string in sf and if I try to select with greater/less than operators, it ignores it and returns a billion rows. Something to do with the fact that CaseNumbers have leading zeros. Could not find a way around that because SOQL does not allow type casting.